+++ src/form/b_input.rs
@@ -1,5 +1,7 @@
 use leptos::*;
 
+use crate::EventFn;
+
 #[component]
 pub fn BInput(
     #[prop(optional, into)] class: Option<&'static str>,
@@ -8,6 +10,8 @@ pub fn BInput(
     #[prop(optional)] name: Option<&'static str>,
     #[prop(optional)] placeholder: Option<&'static str>,
     #[prop(optional, into)] value: MaybeSignal<String>,
+    #[prop(optional, into)] on_change: Option<EventFn>,
+    #[prop(optional, into)] on_input: Option<EventFn>,
     #[prop(attrs, optional)] attributes: Vec<(&'static str, Attribute)>,
 ) -> impl IntoView {
     let mut b_input = view! {
@@ -19,7 +23,9 @@ pub fn BInput(
             placeholder=placeholder
             value=value
         />
-    };
+    }
+    .optional_event(ev::change, on_change.map(EventFn::into_inner))
+    .optional_event(ev::input, on_input.map(EventFn::into_inner));
 
     for (attr_name, attr_value) in attributes {
         b_input = b_input.attr(attr_name, attr_value);