+++ src/form/b_password_field.rs
@@ -4,7 +4,7 @@ use super::{BControl, BField, BHelp, BLabel};
 
 #[component]
 pub fn BPasswordField(
-    #[prop(optional, into)] error: Option<MaybeSignal<Option<String>>>,
+    #[prop(optional, into)] error: MaybeSignal<Option<String>>,
     #[prop(optional)] id: Option<&'static str>,
     #[prop(optional)] label: Option<&'static str>,
     #[prop(optional)] name: Option<&'static str>,
@@ -15,12 +15,7 @@ pub fn BPasswordField(
     let is_visible = create_rw_signal(false);
 
     create_effect(move |_| {
-        error_text.set(
-            error
-                .clone()
-                .and_then(|e| e.get())
-                .map(|e| e.trim().to_owned()),
-        );
+        error_text.set(error.get().map(|e| e.trim().to_owned()));
     });
 
     let has_error = move || error_text.get().is_some();